[김평화] - 주소록 완성

+++ s20240130103/src/main/java/com/blackberry/s20240130103/kph/model/KphUsers.java
@@ -17,6 +17,6 @@ public class KphUsers {
 	private String user_date;
 	
 	// 조회용 테이블
-	private int user_score;
+	private double user_score;
 	
 }
